Remove nb_streams as it is not required anymore as it can be got from the size of the vector

+++ src/c-wrapper/internal/c-sal-media-description.h
@@ -62,7 +62,7 @@ LINPHONE_PUBLIC SalMediaDescription *_linphone_call_get_result_desc (const Linph
 LINPHONE_PUBLIC SalStreamDescription * sal_media_description_get_stream_idx(SalMediaDescription *md, unsigned int idx);
 LINPHONE_PUBLIC SalStreamDescription *sal_media_description_find_stream(SalMediaDescription *md, SalMediaProto proto, SalStreamType type);
 LINPHONE_PUBLIC const char * sal_media_description_get_address(SalMediaDescription *md);
-LINPHONE_PUBLIC int sal_media_description_get_nb_streams(SalMediaDescription *md);
+LINPHONE_PUBLIC size_t sal_media_description_get_nb_streams(SalMediaDescription *md);
 #ifdef __cplusplus
 }
 #endif
